ILSwiss/requirements.txt
absl-py==0.11.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.21
dataclasses
future==0.18.2
glfw==2.0.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
gt==0.3.14
gtimer==1.0.0b5
gym==0.17.3
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.6.0
joblib==0.17.0
kiwisolver==1.3.0
labmaze==1.0.3
lockfile==0.12.2
lxml==4.6.2
Markdown==3.3.4
matplotlib==3.3.2
mujoco-py==1.50.1.68
numpy==1.19.3
oauthlib==3.1.1
olefile==0.46
pandas==1.1.3
path==15.0.0
Pillow
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.22
scipy==1.5.3
seaborn==0.11.0
serializable==0.2.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.0
tqdm==4.51.0
typechecks==0.1.0
typing-extensions
urllib3==1.26.2
pettingzoo
